pyproject.toml
[tool.poetry]
name = "topo-processor"
version = "0.1.0"
description = ""
authors = ["Blayne Chard <bchard@linz.govt.nz>", "Paul Fouquet <pfouquet@linz.govt.nz>", "Megan Davidson <mdavidson@linz.govt.nz>"]

[tool.poetry.dependencies]
python = "^3.8"
linz-logger= "^0.5.0"
py-multihash = "^2.0.1"
rasterio = "^1.2.10"
click = "^8.0.3"
boto3 = "^1.20.10"
python-ulid = "^1.0.2"
fsspec = {extras = ["s3"], version = "^2021.11.0"}
pystac = "^1.2.0"
jsonschema = "^4.2.1"
jsonschema-rs = "^0.13.1"
Shapely = "^1.8.0"
requests = "^2.26.0"
aiohttp = "^3.8.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
isort = "^5.10.1"
pylint = "^2.11.1"
pytest = "^6.2.4"
rope = "^0.21.1"
pytest-cov = "^3.0.0"
pytest-mock = "^3.6.1"
mypy = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 127

[tool.isort]
line_length = 127
case_sensitive = true
profile = "black"

[tool.mypy]
show_error_codes = true
strict = true

[[tool.mypy.overrides]]
module = [
    "linz_logger",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow",
]
